+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: hyejeong <hyejeong@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 13:17:58 by hyejeong          #+#    #+#             */
-/*   Updated: 2022/11/21 15:52:06 by hyejeong         ###   ########.fr       */
+/*   Updated: 2022/11/22 12:37:02 by hyejeong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,8 @@ int		toupper(int c);
 int		ft_strlen(char *str);
 
 void	*ft_memcpy(void *dest, const void *src, size_t n);
-void	*ft_bzero(void *s, unsigned int n);
-void	*ft_memset(void *s, int c, unsigned int n);
+void	*ft_bzero(void *s, size_t n);
+void	*ft_memset(void *s, int c, size_t n);
 void	ft_putchar_fd(char c, int fd);
 void	ft_putendl_fd(char *s, int fd);
 
@@ -37,7 +37,7 @@ char	**ft_split(char const *s, char c);
 char	*strchr(const char *s, int c);
 char	*strdup(const char *s);
 char	*ft_strjoin(char const *s1, char const *s2);
-char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
+char	*ft_strmapi(char const *s, char (*f)(size_t, char));
 
 size_t	strlcpy(char *dst, const char *str, size_t size);
 #endif
